+++ posthog/api/action.py
@@ -23,7 +23,7 @@ from posthog.event_usage import report_user_action
 from posthog.models import Action, ActionStep, Filter, Person
 from posthog.models.action.util import format_action_filter
 from posthog.permissions import ProjectMembershipNecessaryPermissions, TeamMemberAccessPermission
-from posthog.queries.trends.person import TrendsActors
+from posthog.queries.trends.trends_actors import TrendsActors
 
 from .forbid_destroy_model import ForbidDestroyModel
 from .person import get_person_name
